chore(backend): move payment pointer key files
