+++ packages/backend/src/app.ts
@@ -30,7 +30,7 @@ import { AuthService } from './open_payments/auth/service'
 import { RatesService } from './rates/service'
 import { SPSPRoutes } from './spsp/routes'
 import { IncomingPaymentRoutes } from './open_payments/payment/incoming/routes'
-import { PaymentPointerKeyRoutes } from './paymentPointerKey/routes'
+import { PaymentPointerKeyRoutes } from './open_payments/payment_pointer/key/routes'
 import { PaymentPointerRoutes } from './open_payments/payment_pointer/routes'
 import { IncomingPaymentService } from './open_payments/payment/incoming/service'
 import { StreamServer } from '@interledger/stream-receiver'
@@ -46,7 +46,7 @@ import { SessionService } from './session/service'
 import { addDirectivesToSchema } from './graphql/directives'
 import { Session } from './session/util'
 import { createValidatorMiddleware, HttpMethod, isHttpMethod } from 'openapi'
-import { PaymentPointerKeyService } from './paymentPointerKey/service'
+import { PaymentPointerKeyService } from './open_payments/payment_pointer/key/service'
 import { AuthenticatedClient } from 'open-payments'
 
 export interface AppContextData {
