+++ packages/backend/src/open_payments/payment_pointer/key/model.ts
@@ -1,4 +1,4 @@
-import { BaseModel } from '../shared/baseModel'
+import { BaseModel } from '../../../shared/baseModel'
 
 import { JWK } from 'http-signature-utils'
 import { Pojo } from 'objection'
