+++ packages/backend/src/open_payments/payment_pointer/key/routes.test.ts
@@ -3,16 +3,16 @@ import { Knex } from 'knex'
 import { generateJwk, JWK } from 'http-signature-utils'
 import { v4 as uuid } from 'uuid'
 
-import { createContext } from '../tests/context'
-import { createTestApp, TestContainer } from '../tests/app'
-import { Config } from '../config/app'
+import { createContext } from '../../../tests/context'
+import { createTestApp, TestContainer } from '../../../tests/app'
+import { Config } from '../../../config/app'
 import { IocContract } from '@adonisjs/fold'
-import { initIocContainer } from '..'
-import { AppServices, PaymentPointerContext } from '../app'
-import { truncateTables } from '../tests/tableManager'
+import { initIocContainer } from '../../..'
+import { AppServices, PaymentPointerContext } from '../../../app'
+import { truncateTables } from '../../../tests/tableManager'
 import { PaymentPointerKeyRoutes } from './routes'
 import { PaymentPointerKeyService } from './service'
-import { createPaymentPointer } from '../tests/paymentPointer'
+import { createPaymentPointer } from '../../../tests/paymentPointer'
 
 const TEST_KEY: JWK = {
   kid: uuid(),
